Integrate some rpp stuff. We should decide on some correct integration path here, because now changes that gunnar made to the tp2 branch are conflicting with changes roberto made to main. In this, I preferred Roberto's change, since Gunnar's change was only to kick out some compiler warnings.
[git-p4: depot-paths = "//depot/qtjambi/main/": change = 234345]

+++ generator/parser/rpp/preprocessor.cpp
@@ -113,7 +113,7 @@ QStringList Preprocessor::macroNames() const
     pp_environment::const_iterator it = d->env.first_macro();
     while (it != d->env.last_macro()) {
         const pp_macro *m = *it;
-        macros += QString::fromLatin1(m->name->begin(), m->name->size());
+        macros += QString::fromLatin1(m->name->begin(), (int) m->name->size());
         ++it;
     }
 
@@ -128,12 +128,12 @@ QList<Preprocessor::MacroItem> Preprocessor::macros() const
     while (it != d->env.last_macro()) {
         const pp_macro *m = *it;
         MacroItem item;
-        item.name = QString::fromLatin1(m->name->begin(), m->name->size());
+        item.name = QString::fromLatin1(m->name->begin(), (int) m->name->size());
         item.definition = QString::fromLatin1(m->definition->begin(),
-                                              m->definition->size());
+                                              (int) m->definition->size());
         for (size_t i = 0; i < m->formals.size(); ++i) {
             item.parameters += QString::fromLatin1(m->formals[i]->begin(),
-                    m->formals[i]->size());
+                    (int) m->formals[i]->size());
         }
         item.isFunctionLike = m->function_like;
